requirements.txt
streamlit
torch==2.5.1
https://data.pyg.org/whl/torch-2.5.0%2Bcpu/torch_scatter-2.1.2%2Bpt25cpu-cp311-cp311-linux_x86_64.whl
https://data.pyg.org/whl/torch-2.5.0%2Bcpu/torch_sparse-0.6.18%2Bpt25cpu-cp311-cp311-linux_x86_64.whl
torch-geometric
pandas
numpy
easyocr
langdetect==1.0.9
deep-translator
sentence-transformers
scikit-learn
python-docx
pdf2image
Pillow
requests